Added arg for which i2c device to use
