+++ Cargo.toml
@@ -19,15 +19,20 @@ embassy-time = ["dep:embassy-time"]
 sync = ["dep:embedded-hal"]
 
 [dependencies]
-embedded-hal-async = "1.0.0"
-embedded-hal = { version= "1.0.0", optional = true}
-embassy-time = { version = "0.4.0", optional = true }
 defmt = "0.3.8"
+embedded-hal-async = "1.0.0"
+embedded-hal = { version = "1.0.0", optional = true }
+embassy-time = { version = "0.4.0", optional = true }
 libm = "0.2.8"
-uom = { version = "0.36.0", default-features = false, features = ["f32", "si", "autoconvert"] }
+uom = { version = "0.36.0", default-features = false, features = [
+    "f32",
+    "si",
+    "autoconvert",
+] }
 
 [dev-dependencies]
-defmt = { version = "0.3.8", features = [ "unstable-test" ] }
+clap = { version = "4.5.26", features = ["derive"] }
+defmt = { version = "0.3.8", features = ["unstable-test"] }
 embedded-hal-mock = { version = "0.11.0", default-features = false, features = [
     "eh1",
     "embedded-hal-async",
